pyproject.toml
[project]
name = "aiagentplatformpocfinal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.2.6",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langgraph-prebuilt>=1.0.7",
    "python-dotenv>=1.2.1",
]
